requirements.txt
coverage==4.2
Django==1.10.3
django-nose==1.4.4
djangorestframework==3.5.3
djoser==0.5.1
docopt==0.6.2
drf-nested-routers==0.11.1
flake8==3.2.1
mccabe==0.5.2
nose==1.3.7
pluggy==0.4.0
py==1.4.31
pycodestyle==2.2.0
pyflakes==1.3.0
requests==2.12.1
tox==2.5.0
virtualenv==15.1.0
